Moved back to Pi camera and added external power switch
•	Reverted back to use of Raspberry Pi camera as the PU` Aimetis Dual
Lens USB Camera is broken.
•	Added switching to external power when it is available. This is done
using a relay. Switching from batteries to external power is when the
robot is in the garage and only when the external power is stable.
Switching from external power to batteries is done just before the
robot starts driving out of the garage. This because during the power
switch the robot is powered only by a capacitor so the robot cannot
drive during the switch.
Status of robot on webpage now indicates whether power is coming from
external or from batteries.